\name{CQRADMMCPP}
\alias{CQRADMMCPP}
\title{Composite Quantile regression (cqr) use Alternating Direction Method of Multipliers (ADMM) algorithm core computational part}
\description{
   Composite quantile regression (cqr) find the estimated coefficient which minimize the absolute error for various quantile level.
   The problem is well suited to distributed convex optimization and is based on Alternating Direction Method of Multipliers (ADMM) algorithm .
}

\keyword{ADMM}
