\name{CQRCDCPP}
\alias{CQRCDCPP}
\title{Composite Quantile Regression (cqr) use Coordinate Descent (cd) Algorithms core computational part}
\description{
   Composite quantile regression (cqr) find the estimated coefficient which minimize the absolute error for various quantile level.
   The algorithm base on greedy coordinate descent and Edgeworth's for ordinary \eqn{l_1} regression. 
}

\keyword{CD}
