\name{CQRMMCPP}
\alias{CQRMMCPP}
\title{Composite Quantile Regression (cqr) use Majorize and Minimize (mm) Algorithm core computational part}
\description{
   Composite quantile regression find the estimated coefficient which minimize the absolute error for various quantile level.
   The algorithm majorizing the objective function by a quadratic function followed by minimizing that quadratic.
}

\keyword{MM}
