\name{CQRPADMMCPP}
\alias{CQRPADMMCPP}
\title{Composite Quantile Regression (cqr) with Adaptive Lasso Penalty (lasso) use Alternating Direction Method of Multipliers (ADMM) algorithm core computational part}
\description{
   The adaptive lasso parameter base on the estimated coefficient without penalty function.
   Composite quantile regression find the estimated coefficient which minimize the absolute error for various quantile level.
   The problem is well suited to distributed convex optimization and is based on Alternating Direction Method of Multipliers (ADMM) algorithm .
}

\keyword{ADMM}
