\name{CQRPCDCPP}
\alias{CQRPCDCPP}
\title{Composite Quantile Regression (cqr) with Adaptive Lasso Penalty (lasso) use Coordinate Descent (cd) Algorithms core computational part}
\description{
    The adaptive lasso parameter base on the estimated coefficient without penalty function.
    Composite quantile regression find the estimated coefficient which minimize the absolute error for various quantile level.
    The algorithm base on greedy coordinate descent and Edgeworth's for ordinary \eqn{l_1} regression.
}

\keyword{CD}
