\name{CQRPMMCPP}
\alias{CQRPMMCPP}
\title{Composite Quantile Regression (cqr) with Adaptive Lasso Penalty (lasso) use Majorize and Minimize (mm) Algorithm core computational part}
\description{
   The adaptive lasso penalty parameter base on the estimated coefficient without penalty function.
   Composite quantile regression find the estimated coefficient which minimize the absolute error for various quantile level.
   The algorithm majorizing the objective function by a quadratic function followed by minimizing that quadratic. 
}

\keyword{MM}
