\name{QRADMMCPP}
\alias{QRADMMCPP}
\title{Quantile Regression (QR) use Alternating Direction Method of Multipliers (ADMM) algorithm core computational part}
\description{
   The problem is well suited to distributed convex optimization and is based on Alternating Direction Method of Multipliers (ADMM) algorithm .
}

\keyword{ADMM}
