\name{QRPADMMCPP}
\alias{QRPADMMCPP}
\title{Quantile Regression (QR) with Adaptive Lasso Penalty (lasso) use Alternating Direction
Method of Multipliers (ADMM) algorithm core computational part}
\description{
   The adaptive lasso parameter base on the estimated coefficient without penalty function.
    The problem is well suited to distributed convex optimization and is based on Alternating Direction Method of Multipliers (ADMM) algorithm .
}
\keyword{ADMM}
