\name{QRPCDCPP}
\alias{QRPCDCPP}
\title{Quantile Regression (QR) with Adaptive Lasso Penalty (lasso) use Coordinate Descent (cd)  Algorithms core computational part}
\description{
   The adaptive lasso parameter base on the estimated coefficient without penalty function.
   The algorithm base on greedy coordinate descent and Edgeworth's for ordinary \eqn{l_1}    regression. As explored by Tong Tong Wu and Kenneth Lange.
}
\keyword{CD}
