\name{QRPMMCPP}
\alias{QRPMMCPP}
\title{Quantile Regression (QR) with Adaptive Lasso Penalty (lasso) use Majorize and Minimize (mm) algorithm core computational part}
\description{
   The adaptive lasso parameter base on the estimated coefficient without penalty function.
   The algorithm majorizing the objective function by a quadratic function followed by minimizing that quadratic.
}

\keyword{MM}
